Missing file in previous commit
Former-commit-id: eeda5f19f45a1d080a5a14013aa80e934a848260 [formerly 158db20ddfebd99ba467a3b99c7da8d2059a4b43]
Former-commit-id: 7a28598b3d532b66a14238e7a0095093c52c0d49

+++ av/NativeAudioNode.h
@@ -208,8 +208,8 @@ class NativeAudioNodeDelay : public NativeAudioNode
         };
 
         double delay;
-        int wet;
-        int dry;
+        double wet;
+        double dry;
         float **buffers;
 
         virtual bool process();
